Changed algorithm

+++ f.hpp
@@ -1,6 +1,6 @@
 #ifndef F_HPP_
 #define F_HPP_
 extern "C" {
-	void f( unsigned char *pBuffer, int W, int H, int A, int B, int C, int D, int S );
+	void f( unsigned char *pBuffer, int W, int H, float A, float B, float C, float D, float S );
 }
 #endif // F_HPP_
